requirements.txt
openai-agents
python-dotenv
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
reportlab
sendgrid
pydantic
reportlab